requirements.txt
pytest==7.4.2
requests==2.32.0
pytest-html==3.2.0
playwright~=1.49.1
python-dotenv==1.0.1